add a new parameter @source_as_temp_table to improve merge script performance
we have a table with about 5k records need to generate a merge script,
while I choose to not delete when not matched SQL Server would complaint
that 'query processor ran out of internal resources'.

I suppose as the merge source is all in memory and causing the trouble,
I moved all source into a temporary table and then use that as the
source for the merge and the script ran successful without any problem.

This doesn't mean it will run faster, in fact it might even slow it down
on limited disk resource servers, but at least it would run as SQL
Server won't consume too much memory to form the source table which
might causing the error.

ACTUAL ERROR is:

Msg 8623, Level 16, State 1, Line 8
The query processor ran out of internal resources and could not produce
a query plan. This is a rare event and only expected for extremely
complex queries or queries that reference a very large number of tables
or partitions. Please simplify the query. If you believe you have
received this message in error, contact Customer Support Services for
more information.