Added order by primary key
In tables with thousands of records and more, ordering was not consistent. Simplified example: 1, 2, 3, 7, 8, 9, 4, 5, 6...
The problem is when you try to compare two versions of script. Compare tool (e.g. in Visual Studio) detects changes because of the different order when there are actually no changes in raw data.

The solution is to always force ordering by table primary key.