Bugfix: TIMESTAMP column without a default would cause errors when attempting to apply the DEFAULT.
Even though a default isn't present on the TIMESTAMP column SQL Server handles updating the value behind the scenes. Since there's no need to use DEFAULT, and because SQL Server does not allow INSERT or UPDATE of TIMESTAMP/ROWVERSION columns with a specifiied value, I've removed TIMESTAMP columns from the generated MERGE statements.

This also means that @include_timestamp is now pointless and could be removed. I've opted to leave it in and indicate that it's deprecated as a comment in order to preserve backwards compatability.