Exclude computed columns from the MERGE by default
For the primary use case of this project, whereby static data is generated from a populated table, and subsequently used to populate other environments, it does not make sense to include computed columns in the merge statement; given that the structure of the table should be identical between environments, including the computed column spec, including such columns in the merge is in itself redundant. The only use case in which it may make sense to include the computed columns in the merge is in ETL operations (i.e. with the `@include_values` param is set to `0`), where the target table may not fully match the structure of the source.