Don't use the `ALL` keyword to disable constraints if they will only be re-enabled selectively
If the generated MERGE is executed on an environment that has a slightly different constraint configuration (primarily, where there are more constraints in the target env than in the source), then this could result in more constraints being disabled than enabled. The solution is to perform both the disabling and enabling of constraints selectively, in the scenario where the WITH CHECK/WITH NOCHECK/DISABLED status varies between constraints in the source environment.

However, note that if the status of all the source table's constraints is consistent, then the `ALL` keyword will continue to be used when disabling/enabling constraints.