Update master.dbo.sp_generate_merge.sql
Added "SET QUOTED_IDENTIFIER ON" on lines 2 and 202.  I was getting an error:

SELECT failed because the following SET options have incorrect settings: 'QUOTED_IDENTIFIER'. Verify that SET options are correct for use with indexed views and/or indexes on computed columns and/or filtered indexes and/or query notifications and/or XML data type methods and/or spatial index operations.

I found an apparent solution at https://dba.stackexchange.com/a/233473 and then further down someone points out that you need to add that to the procedure as well.  It can be added before the procedure create, which is how I've done it here, and I also added it inside the procedure.  I think it may be only needed within the proc, but I don't believe it'll hurt..